backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.1
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.3.0
python-dotenv==1.0.0
